+++ Guia 1/ex_12.cpp
@@ -4,9 +4,9 @@ using namespace std;
 int main ()
 {
     int huevos_cliente;
-    int docena = 1000;
-    int huevo_suelto = 120;
-    int total_pagar;
+    float docena = 1000;
+    float huevo_suelto = 120;
+    float total_pagar;
     int docenas_pedidas;
     int huevos_extra;
